Main: Mesh - add UserObjectBindings & merge PrefabLoader + PrefabFactory
we can now use UserObjectBindings to pass
the procedural parameters to the Listener
