+++ OgreMain/src/OgrePrefabFactory.cpp
@@ -30,28 +30,51 @@ THE SOFTWARE.
 #include "OgrePrefabFactory.h"
 
 namespace Ogre {
-    //---------------------------------------------------------------------
-    bool PrefabFactory::createPrefab(Mesh* mesh)
+    //-----------------------------------------------------------------------
+    void PrefabFactory::loadResource(Resource* res)
     {
-        const String& resourceName = mesh->getName();
+        Mesh* msh = static_cast<Mesh*>(res);
 
+        // attempt to create a prefab mesh
+        const String& resourceName = msh->getName();
         if(resourceName == "Prefab_Plane")
         {
-            createPlane(mesh);
-            return true;
+            createPlane(msh);
+            return;
         }
         else if(resourceName == "Prefab_Cube")
         {
-            createCube(mesh);
-            return true;
+            createCube(msh);
+            return;
         }
         else if(resourceName == "Prefab_Sphere")
         {
-            createSphere(mesh);
-            return true;
+            createSphere(msh);
+            return;
         }
 
-        return false;
+        // the mesh was not a prefab.. grab build parameters
+        auto any = msh->getUserObjectBindings().getUserAny("_MeshBuildParams");
+
+        if (!any.has_value())
+            OGRE_EXCEPT(Exception::ERR_ITEM_NOT_FOUND, "No prefab parameters in " + res->getName());
+
+        auto params = any_cast<MeshBuildParams>(any);
+
+        switch(params.type)
+        {
+        case MBT_PLANE:
+            loadManualPlane(msh, params);
+            return;
+        case MBT_CURVED_ILLUSION_PLANE:
+            loadManualCurvedIllusionPlane(msh, params);
+            return;
+        case MBT_CURVED_PLANE:
+            loadManualCurvedPlane(msh, params);
+            return;
+        }
+
+        OgreAssertDbg(false, "unknown prefab type");
     }
     //---------------------------------------------------------------------
     void PrefabFactory::createPlane(Mesh* mesh)
@@ -354,5 +377,517 @@ namespace Ogre {
             Vector3(SPHERE_RADIUS, SPHERE_RADIUS, SPHERE_RADIUS) ), false );
 
         mesh->_setBoundingSphereRadius(SPHERE_RADIUS);
+    }
+    //-----------------------------------------------------------------------
+    static void tesselate2DMesh(SubMesh* sm, unsigned short meshWidth, unsigned short meshHeight, bool doubleSided,
+                                HardwareBuffer::Usage indexBufferUsage, bool indexShadowBuffer)
+    {
+        // The mesh is built, just make a list of indexes to spit out the triangles
+        unsigned short vInc, v, iterations;
+
+        if (doubleSided)
+        {
+            iterations = 2;
+            vInc = 1;
+            v = 0; // Start with front
+        }
+        else
+        {
+            iterations = 1;
+            vInc = 1;
+            v = 0;
+        }
+
+        // Allocate memory for faces
+        // Num faces, width*height*2 (2 tris per square), index count is * 3 on top
+        sm->indexData->indexCount = (meshWidth-1) * (meshHeight-1) * 2 * iterations * 3;
+        sm->indexData->indexBuffer = HardwareBufferManager::getSingleton().
+            createIndexBuffer(HardwareIndexBuffer::IT_16BIT,
+            sm->indexData->indexCount, indexBufferUsage, indexShadowBuffer);
+
+        unsigned short v1, v2, v3;
+        //bool firstTri = true;
+        HardwareIndexBufferSharedPtr ibuf = sm->indexData->indexBuffer;
+        // Lock the whole buffer
+        HardwareBufferLockGuard ibufLock(ibuf, HardwareBuffer::HBL_DISCARD);
+        unsigned short* pIndexes = static_cast<unsigned short*>(ibufLock.pData);
+
+        while (iterations--)
+        {
+            // Make tris in a zigzag pattern (compatible with strips)
+            unsigned short u = 0;
+            unsigned short uInc = 1; // Start with moving +u
+            unsigned short vCount = meshHeight - 1;
+
+            while (vCount--)
+            {
+                unsigned short uCount = meshWidth - 1;
+                while (uCount--)
+                {
+                    // First Tri in cell
+                    // -----------------
+                    v1 = ((v + vInc) * meshWidth) + u;
+                    v2 = (v * meshWidth) + u;
+                    v3 = ((v + vInc) * meshWidth) + (u + uInc);
+                    // Output indexes
+                    *pIndexes++ = v1;
+                    *pIndexes++ = v2;
+                    *pIndexes++ = v3;
+                    // Second Tri in cell
+                    // ------------------
+                    v1 = ((v + vInc) * meshWidth) + (u + uInc);
+                    v2 = (v * meshWidth) + u;
+                    v3 = (v * meshWidth) + (u + uInc);
+                    // Output indexes
+                    *pIndexes++ = v1;
+                    *pIndexes++ = v2;
+                    *pIndexes++ = v3;
+
+                    // Next column
+                    u += uInc;
+                }
+                // Next row
+                v += vInc;
+                u = 0;
+
+
+            }
+
+            // Reverse vInc for double sided
+            v = meshHeight - 1;
+            vInc = -vInc;
+
+        }
+    }
+    //-----------------------------------------------------------------------
+    void PrefabFactory::loadManualPlane(Mesh* pMesh, MeshBuildParams& params)
+    {
+        if ((params.xsegments + 1) * (params.ysegments + 1) > 65536)
+            OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS,
+                "Plane tessellation is too high, must generate max 65536 vertices");
+        SubMesh *pSub = pMesh->createSubMesh();
+
+        // Set up vertex data
+        // Use a single shared buffer
+        pMesh->sharedVertexData = OGRE_NEW VertexData();
+        VertexData* vertexData = pMesh->sharedVertexData;
+        // Set up Vertex Declaration
+        VertexDeclaration* vertexDecl = vertexData->vertexDeclaration;
+        size_t currOffset = 0;
+        // We always need positions
+        currOffset += vertexDecl->addElement(0, currOffset, VET_FLOAT3, VES_POSITION).getSize();
+        // Optional normals
+        if(params.normals)
+        {
+            currOffset += vertexDecl->addElement(0, currOffset, VET_FLOAT3, VES_NORMAL).getSize();
+        }
+
+        for (unsigned short i = 0; i < params.numTexCoordSets; ++i)
+        {
+            // Assumes 2D texture coords
+            currOffset += vertexDecl->addElement(0, currOffset, VET_FLOAT2, VES_TEXTURE_COORDINATES, i).getSize();
+        }
+
+        vertexData->vertexCount = (params.xsegments + 1) * (params.ysegments + 1);
+
+        // Allocate vertex buffer
+        HardwareVertexBufferSharedPtr vbuf =
+            pMesh->getHardwareBufferManager()->createVertexBuffer(
+                vertexDecl->getVertexSize(0), vertexData->vertexCount,
+                params.vertexBufferUsage, params.vertexShadowBuffer);
+
+        // Set up the binding (one source only)
+        VertexBufferBinding* binding = vertexData->vertexBufferBinding;
+        binding->setBinding(0, vbuf);
+
+        // Work out the transform required
+        // Default orientation of plane is normal along +z, distance 0
+        Affine3 xlate, xform, rot;
+        Matrix3 rot3;
+        xlate = rot = Affine3::IDENTITY;
+        // Determine axes
+        Vector3 zAxis, yAxis, xAxis;
+        zAxis = params.plane.normal;
+        zAxis.normalise();
+        yAxis = params.upVector;
+        yAxis.normalise();
+        xAxis = yAxis.crossProduct(zAxis);
+        if (xAxis.squaredLength() == 0)
+        {
+            //upVector must be wrong
+            OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS, "The upVector you supplied is parallel to the plane normal, so is not valid.");
+        }
+        xAxis.normalise();
+
+        rot3.FromAxes(xAxis, yAxis, zAxis);
+        rot = rot3;
+
+        // Set up standard transform from origin
+        xlate.setTrans(params.plane.normal * -params.plane.d);
+
+        // concatenate
+        xform = xlate * rot;
+
+        // Generate vertex data
+        // Lock the whole buffer
+        HardwareBufferLockGuard vbufLock(vbuf, HardwareBuffer::HBL_DISCARD);
+        float* pReal = static_cast<float*>(vbufLock.pData);
+        Real xSpace = params.width / params.xsegments;
+        Real ySpace = params.height / params.ysegments;
+        Real halfWidth = params.width / 2;
+        Real halfHeight = params.height / 2;
+        Real xTex = (1.0f * params.xTile) / params.xsegments;
+        Real yTex = (1.0f * params.yTile) / params.ysegments;
+        Vector3 vec;
+        AxisAlignedBox aabb;
+
+        for (int y = 0; y < params.ysegments + 1; ++y)
+        {
+            for (int x = 0; x < params.xsegments + 1; ++x)
+            {
+                // Work out centered on origin
+                vec.x = (x * xSpace) - halfWidth;
+                vec.y = (y * ySpace) - halfHeight;
+                vec.z = 0.0f;
+                // Transform by orientation and distance
+                vec = xform * vec;
+                // Assign to geometry
+                *pReal++ = vec.x;
+                *pReal++ = vec.y;
+                *pReal++ = vec.z;
+
+                // Build bounds as we go
+                aabb.merge(vec);
+
+                if (params.normals)
+                {
+                    // Default normal is along unit Z
+                    vec = Vector3::UNIT_Z;
+                    // Rotate
+                    vec = rot * vec;
+
+                    *pReal++ = vec.x;
+                    *pReal++ = vec.y;
+                    *pReal++ = vec.z;
+                }
+
+                for (unsigned short i = 0; i < params.numTexCoordSets; ++i)
+                {
+                    *pReal++ = x * xTex;
+                    *pReal++ = 1 - (y * yTex);
+                }
+
+
+            } // x
+        } // y
+
+        // Unlock
+        vbufLock.unlock();
+        // Generate face list
+        pSub->useSharedVertices = true;
+        tesselate2DMesh(pSub, params.xsegments + 1, params.ysegments + 1, false,
+            params.indexBufferUsage, params.indexShadowBuffer);
+
+        pMesh->_setBounds(aabb, true);
+    }
+    //-----------------------------------------------------------------------
+    void PrefabFactory::loadManualCurvedPlane(Mesh* pMesh, MeshBuildParams& params)
+    {
+        if ((params.xsegments + 1) * (params.ysegments + 1) > 65536)
+            OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS,
+                "Plane tessellation is too high, must generate max 65536 vertices");
+        SubMesh *pSub = pMesh->createSubMesh();
+
+        // Set options
+        pMesh->sharedVertexData = OGRE_NEW VertexData();
+        pMesh->sharedVertexData->vertexStart = 0;
+        VertexBufferBinding* bind = pMesh->sharedVertexData->vertexBufferBinding;
+        VertexDeclaration* decl = pMesh->sharedVertexData->vertexDeclaration;
+
+        pMesh->sharedVertexData->vertexCount = (params.xsegments + 1) * (params.ysegments + 1);
+
+        size_t offset = 0;
+        offset += decl->addElement(0, offset, VET_FLOAT3, VES_POSITION).getSize();
+        if (params.normals)
+        {
+            offset += decl->addElement(0, 0, VET_FLOAT3, VES_NORMAL).getSize();
+        }
+
+        for (unsigned short i = 0; i < params.numTexCoordSets; ++i)
+        {
+            offset += decl->addElement(0, offset, VET_FLOAT2, VES_TEXTURE_COORDINATES, i).getSize();
+        }
+
+
+        // Allocate memory
+        HardwareVertexBufferSharedPtr vbuf =
+            pMesh->getHardwareBufferManager()->createVertexBuffer(
+                offset, pMesh->sharedVertexData->vertexCount,
+                params.vertexBufferUsage, params.vertexShadowBuffer);
+        bind->setBinding(0, vbuf);
+
+        // Work out the transform required
+        // Default orientation of plane is normal along +z, distance 0
+        Affine3 xlate, xform, rot;
+        Matrix3 rot3;
+        xlate = rot = Affine3::IDENTITY;
+        // Determine axes
+        Vector3 zAxis, yAxis, xAxis;
+        zAxis = params.plane.normal;
+        zAxis.normalise();
+        yAxis = params.upVector;
+        yAxis.normalise();
+        xAxis = yAxis.crossProduct(zAxis);
+        if (xAxis.squaredLength() == 0)
+        {
+            //upVector must be wrong
+            OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS, "The upVector you supplied is parallel to the plane normal, so is not valid.");
+        }
+        xAxis.normalise();
+
+        rot3.FromAxes(xAxis, yAxis, zAxis);
+        rot = rot3;
+
+        // Set up standard transform from origin
+        xlate.setTrans(params.plane.normal * -params.plane.d);
+
+        // concatenate
+        xform = xlate * rot;
+
+        // Generate vertex data
+        HardwareBufferLockGuard vbufLock(vbuf, HardwareBuffer::HBL_DISCARD);
+        float* pFloat = static_cast<float*>(vbufLock.pData);
+        Real xSpace = params.width / params.xsegments;
+        Real ySpace = params.height / params.ysegments;
+        Real halfWidth = params.width / 2;
+        Real halfHeight = params.height / 2;
+        Real xTex = (1.0f * params.xTile) / params.xsegments;
+        Real yTex = (1.0f * params.yTile) / params.ysegments;
+        Vector3 vec;
+
+        AxisAlignedBox aabb;
+
+        Real diff_x, diff_y, dist;
+
+        for (int y = 0; y < params.ysegments + 1; ++y)
+        {
+            for (int x = 0; x < params.xsegments + 1; ++x)
+            {
+                // Work out centered on origin
+                vec.x = (x * xSpace) - halfWidth;
+                vec.y = (y * ySpace) - halfHeight;
+
+                // Here's where curved plane is different from standard plane.  Amazing, I know.
+                diff_x = (x - ((params.xsegments) / 2)) / static_cast<Real>((params.xsegments));
+                diff_y = (y - ((params.ysegments) / 2)) / static_cast<Real>((params.ysegments));
+                dist = std::sqrt(diff_x*diff_x + diff_y * diff_y );
+                vec.z = (-std::sin((1-dist) * (Math::PI/2)) * params.curvature) + params.curvature;
+
+                // Transform by orientation and distance
+                Vector3 pos = xform * vec;
+                // Assign to geometry
+                *pFloat++ = pos.x;
+                *pFloat++ = pos.y;
+                *pFloat++ = pos.z;
+
+                // Record bounds
+                aabb.merge(vec);
+
+                if (params.normals)
+                {
+                    // This part is kinda 'wrong' for curved planes... but curved planes are
+                    //   very valuable outside sky planes, which don't typically need normals
+                    //   so I'm not going to mess with it for now.
+
+                    // Default normal is along unit Z
+                    //vec = Vector3::UNIT_Z;
+                    // Rotate
+                    vec = rot * vec;
+                    vec.normalise();
+
+                    *pFloat++ = vec.x;
+                    *pFloat++ = vec.y;
+                    *pFloat++ = vec.z;
+                }
+
+                for (unsigned short i = 0; i < params.numTexCoordSets; ++i)
+                {
+                    *pFloat++ = x * xTex;
+                    *pFloat++ = 1 - (y * yTex);
+                }
+
+            } // x
+        } // y
+        vbufLock.unlock();
+
+        // Generate face list
+        tesselate2DMesh(pSub, params.xsegments + 1, params.ysegments + 1,
+            false, params.indexBufferUsage, params.indexShadowBuffer);
+
+        pMesh->_setBounds(aabb, true);
+    }
+    //-----------------------------------------------------------------------
+    void PrefabFactory::loadManualCurvedIllusionPlane(Mesh* pMesh, MeshBuildParams& params)
+    {
+        if (params.ySegmentsToKeep == -1) params.ySegmentsToKeep = params.ysegments;
+
+        if ((params.xsegments + 1) * (params.ySegmentsToKeep + 1) > 65536)
+            OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS,
+                "Plane tessellation is too high, must generate max 65536 vertices");
+        SubMesh *pSub = pMesh->createSubMesh();
+
+
+        // Set up vertex data
+        // Use a single shared buffer
+        pMesh->sharedVertexData = OGRE_NEW VertexData();
+        VertexData* vertexData = pMesh->sharedVertexData;
+        // Set up Vertex Declaration
+        VertexDeclaration* vertexDecl = vertexData->vertexDeclaration;
+        size_t currOffset = 0;
+        // We always need positions
+        currOffset += vertexDecl->addElement(0, currOffset, VET_FLOAT3, VES_POSITION).getSize();
+        // Optional normals
+        if(params.normals)
+        {
+            currOffset += vertexDecl->addElement(0, currOffset, VET_FLOAT3, VES_NORMAL).getSize();
+        }
+
+        for (unsigned short i = 0; i < params.numTexCoordSets; ++i)
+        {
+            // Assumes 2D texture coords
+            currOffset += vertexDecl->addElement(0, currOffset, VET_FLOAT2, VES_TEXTURE_COORDINATES, i).getSize();
+        }
+
+        vertexData->vertexCount = (params.xsegments + 1) * (params.ySegmentsToKeep + 1);
+
+        // Allocate vertex buffer
+        HardwareVertexBufferSharedPtr vbuf =
+            pMesh->getHardwareBufferManager()->createVertexBuffer(
+                vertexDecl->getVertexSize(0), vertexData->vertexCount,
+                params.vertexBufferUsage, params.vertexShadowBuffer);
+
+        // Set up the binding (one source only)
+        VertexBufferBinding* binding = vertexData->vertexBufferBinding;
+        binding->setBinding(0, vbuf);
+
+        // Work out the transform required
+        // Default orientation of plane is normal along +z, distance 0
+        Affine3 xlate, xform, rot;
+        Matrix3 rot3;
+        xlate = rot = Affine3::IDENTITY;
+        // Determine axes
+        Vector3 zAxis, yAxis, xAxis;
+        zAxis = params.plane.normal;
+        zAxis.normalise();
+        yAxis = params.upVector;
+        yAxis.normalise();
+        xAxis = yAxis.crossProduct(zAxis);
+        if (xAxis.squaredLength() == 0)
+        {
+            //upVector must be wrong
+            OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS, "The upVector you supplied is parallel to the plane normal, so is not valid.");
+        }
+        xAxis.normalise();
+
+        rot3.FromAxes(xAxis, yAxis, zAxis);
+        rot = rot3;
+
+        // Set up standard transform from origin
+        xlate.setTrans(params.plane.normal * -params.plane.d);
+
+        // concatenate
+        xform = xlate * rot;
+
+        // Generate vertex data
+        // Imagine a large sphere with the camera located near the top
+        // The lower the curvature, the larger the sphere
+        // Use the angle from viewer to the points on the plane
+        // Credit to Aftershock for the general approach
+        Real camPos;      // Camera position relative to sphere center
+
+        // Derive sphere radius
+        Vector3 vertPos;  // position relative to camera
+        Real sphDist;      // Distance from camera to sphere along box vertex vector
+        // Vector3 camToSph; // camera position to sphere
+        Real sphereRadius;// Sphere radius
+        // Actual values irrelevant, it's the relation between sphere radius and camera position that's important
+        const Real SPHERE_RAD = 100.0;
+        const Real CAM_DIST = 5.0;
+
+        sphereRadius = SPHERE_RAD - params.curvature;
+        camPos = sphereRadius - CAM_DIST;
+
+        // Lock the whole buffer
+        HardwareBufferLockGuard vbufLock(vbuf, HardwareBuffer::HBL_DISCARD);
+        float* pFloat = static_cast<float*>(vbufLock.pData);
+        Real xSpace = params.width / params.xsegments;
+        Real ySpace = params.height / params.ysegments;
+        Real halfWidth = params.width / 2;
+        Real halfHeight = params.height / 2;
+        Vector3 vec, norm;
+        AxisAlignedBox aabb;
+
+        for (int y = params.ysegments - params.ySegmentsToKeep; y < params.ysegments + 1; ++y)
+        {
+            for (int x = 0; x < params.xsegments + 1; ++x)
+            {
+                // Work out centered on origin
+                vec.x = (x * xSpace) - halfWidth;
+                vec.y = (y * ySpace) - halfHeight;
+                vec.z = 0.0f;
+                // Transform by orientation and distance
+                vec = xform * vec;
+                // Assign to geometry
+                *pFloat++ = vec.x;
+                *pFloat++ = vec.y;
+                *pFloat++ = vec.z;
+
+                // Build bounds as we go
+                aabb.merge(vec);
+
+                if (params.normals)
+                {
+                    // Default normal is along unit Z
+                    norm = Vector3::UNIT_Z;
+                    // Rotate
+                    norm = params.orientation * norm;
+
+                    *pFloat++ = norm.x;
+                    *pFloat++ = norm.y;
+                    *pFloat++ = norm.z;
+                }
+
+                // Generate texture coords
+                // Normalise position
+                // modify by orientation to return +y up
+                vec = params.orientation.Inverse() * vec;
+                vec.normalise();
+                // Find distance to sphere
+                sphDist = Math::Sqrt(camPos*camPos * (vec.y*vec.y-1.0f) + sphereRadius*sphereRadius) - camPos*vec.y;
+
+                vec.x *= sphDist;
+                vec.z *= sphDist;
+
+                // Use x and y on sphere as texture coordinates, tiled
+                Real s = vec.x * (0.01f * params.xTile);
+                Real t = 1.0f - (vec.z * (0.01f * params.yTile));
+                for (unsigned short i = 0; i < params.numTexCoordSets; ++i)
+                {
+                    *pFloat++ = s;
+                    *pFloat++ = t;
+                }
+
+
+            } // x
+        } // y
+
+        // Unlock
+        vbufLock.unlock();
+        // Generate face list
+        pSub->useSharedVertices = true;
+        tesselate2DMesh(pSub, params.xsegments + 1, params.ySegmentsToKeep + 1, false,
+            params.indexBufferUsage, params.indexShadowBuffer);
+
+        pMesh->_setBounds(aabb, true);
     }
 }
